+++ shm/cargo.hpp
@@ -3,7 +3,6 @@
 #include <memory>
 #include <string>
 
-#include "subjectTime.hpp"
 #include "time.hpp"
 
 class Cargo : public ObserverTime {
