--- shm/delegate.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <cstdint>
-
-class Delegate {
-public:
-    virtual ~Delegate() = default;
-    virtual void payCrew(uint32_t salary) = 0;
-};
