--- shm/observerTime.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-class ObserverTime {
-public:
-    virtual void nextDay() = 0;
-    virtual ~ObserverTime() = default;
-};
